requirements.txt
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
blinker==1.4
celery==5.2.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Flask==2.1.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.3
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.37
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.3.2
Werkzeug==2.1.2
wrapt==1.14.1
